% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_xtrial_data.R
\name{get_xtrial_dgp_parameters}
\alias{get_xtrial_dgp_parameters}
\title{Get cross-trial DGP parameters}
\usage{
get_xtrial_dgp_parameters(dgp_choice = "delta1")
}
\arguments{
\item{dgp_choice}{Character. One of \code{"delta0"}, \code{"delta1"},
\code{"delta2"}, or \code{"delta3"}.}
}
\value{
A list containing scenario-specific DGP parameters.
}
\description{
Get cross-trial DGP parameters
}
\examples{
get_xtrial_dgp_parameters("delta1")

}
