% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_xtrial_data.R
\name{get_tb_dgp_parameters}
\alias{get_tb_dgp_parameters}
\title{Backward-compatible alias for get_xtrial_dgp_parameters()}
\usage{
get_tb_dgp_parameters(...)
}
\description{
Backward-compatible alias for get_xtrial_dgp_parameters()
}
\keyword{internal}
