% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tb_data.R
\name{get_tb_dgp_parameters}
\alias{get_tb_dgp_parameters}
\title{Get TB DGP parameters}
\usage{
get_tb_dgp_parameters(dgp_choice = "delta1")
}
\arguments{
\item{dgp_choice}{Character. One of \code{"delta0"}, \code{"delta1"},
\code{"delta2"}, or \code{"delta3"}.}
}
\value{
A list containing scenario-specific DGP parameters.
}
\description{
Get TB DGP parameters
}
\examples{
get_tb_dgp_parameters("delta1")

}
