% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtrial.R
\name{tb_compare}
\alias{tb_compare}
\title{Backward-compatible alias for xtrial()}
\usage{
tb_compare(...)
}
\value{
An \code{xtrial} object.
}
\description{
Backward-compatible alias for xtrial()
}
\keyword{internal}
