% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_compare.R
\name{tb_compare}
\alias{tb_compare}
\title{Main interface for TB comparison estimators}
\usage{
tb_compare(
  data,
  approach = c("traditional", "tte", "direct", "indirect"),
  estimator = c("ipw", "gcomp", "onestep", "tmle"),
  nboot = 100,
  nuisance_type = c("simple", "flexible")
)
}
\arguments{
\item{data}{Data frame.}

\item{approach}{One of
"traditional",
"observational",
"direct",
"indirect".}

\item{estimator}{One of
"ipw",
"gcomp",
"onestep",
"tmle".}

\item{nboot}{Number of bootstrap samples for bootstrap-based estimators.}

\item{nuisance_type}{Nuisance estimation type. Currently mainly used for EIF/TMLE; one of "simple" or "flexible".}
}
\value{
A tbcompare object.
}
\description{
Main interface for TB comparison estimators
}
